Fix Visual Studio type narrowing warnings.
Visual Studio is far more aggressive about type narrowing than other compilers.
It should be noted that the biggest change of this is that we have
found that zmq_poll takes a system specific long type which means that
our max polling duration on windows is significantly smaller than on other
operating systems. That means on windows we should move to zmq 3 asap.

+++ remus/common/PollingMonitor.cxx
@@ -90,7 +90,7 @@ public:
       }
     else
       {
-      long mil_secs = avg.total_milliseconds() * 0.25;
+      long mil_secs = static_cast<long>(avg.total_milliseconds() * 0.25);
       this->CurrentPollRate = avg + milliseconds(mil_secs);
       if(this->CurrentPollRate > this->MaxTimeOut)
         {
