+++ remus/server/Server.cxx
@@ -304,7 +304,7 @@ bool Server::brokering(Server::SignalHandling sh)
   boost::int64_t timeToCheckForDeadWorkers = 0; //check every 250ms
   while (Thread->isBrokering())
     {
-    zmq::poll(&items[0], 2, monitor.current()*1000);
+    zmq::poll(&items[0], 2, static_cast<long>(monitor.current()*1000));
     timeToCheckForDeadWorkers += monitor.durationOfTheLastPollMilliseconds();
     monitor.pollOccurred();
